requirements.txt
# JAX and related libraries
jax==0.6.0
jaxlib==0.6.0
flax==0.10.6
optax==0.2.4
orbax-checkpoint==0.11.10

# Scientific computing
numpy==2.1.3
scipy==1.15.2
scikit-image==0.25.2

# Image processing
opencv-python==4.11.0.86
tifffile==2025.3.13
mrcfile>=1.4.3

# Deep learning (only for DataLoader)
torch==2.7.0
tensorboard==2.19.0

# Utilities
tqdm==4.67.1
pyyaml==6.0.2
h5py==3.13.0
matplotlib==3.10.1
Pillow==11.1.0